[models] Fix type comparison warnings.

+++ src/models/src/NIMutableTableViewModel.m
@@ -42,7 +42,7 @@
 
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
-- (NSArray *)addObject:(id)object toSection:(NSInteger)sectionIndex {
+- (NSArray *)addObject:(id)object toSection:(NSUInteger)sectionIndex {
   NIDASSERT(sectionIndex >= 0 && sectionIndex < self.sections.count);
   NITableViewModelSection *section = [self.sections objectAtIndex:sectionIndex];
   [section.mutableRows addObject:object];
@@ -63,13 +63,13 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (NSArray *)removeObjectAtIndexPath:(NSIndexPath *)indexPath {
-  NIDASSERT(indexPath.section < self.sections.count);
-  if (indexPath.section >= self.sections.count) {
+  NIDASSERT(indexPath.section < (NSInteger)self.sections.count);
+  if (indexPath.section >= (NSInteger)self.sections.count) {
     return nil;
   }
   NITableViewModelSection* section = [self.sections objectAtIndex:indexPath.section];
-  NIDASSERT(indexPath.row < section.mutableRows.count);
-  if (indexPath.row >= section.mutableRows.count) {
+  NIDASSERT(indexPath.row < (NSInteger)section.mutableRows.count);
+  if (indexPath.row >= (NSInteger)section.mutableRows.count) {
     return nil;
   }
   [section.mutableRows removeObjectAtIndex:indexPath.row];
@@ -86,7 +86,7 @@
 
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
-- (NSIndexSet *)insertSectionWithTitle:(NSString *)title atIndex:(NSInteger)index {
+- (NSIndexSet *)insertSectionWithTitle:(NSString *)title atIndex:(NSUInteger)index {
   NITableViewModelSection* section = [self _insertSectionAtIndex:index];
   section.headerTitle = title;
   return [NSIndexSet indexSetWithIndex:index];
@@ -112,7 +112,7 @@
 
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
-- (NITableViewModelSection *)_insertSectionAtIndex:(NSInteger)index {
+- (NITableViewModelSection *)_insertSectionAtIndex:(NSUInteger)index {
   if (nil == self.sections) {
     self.sections = [NSMutableArray array];
   }
